+++ test/1150-commands_MPAddAreaBorderTHIDCommand_test.dart
@@ -15,7 +15,7 @@ import 'th_test_aux.dart';
 class FakePathProviderPlatform extends PathProviderPlatform {
   @override
   Future<String?> getApplicationDocumentsPath() async {
-    return '/tmp'; // or any fake path
+    return '/tmp';
   }
 }
 
@@ -25,9 +25,7 @@ void main() {
   final MPLocator mpLocator = MPLocator();
   group('command: MPAddAreaBorderTHIDCommand', () {
     setUp(() {
-      // Provide localizations to avoid null access in updateUndoRedoStatus()
       mpLocator.appLocalizations = AppLocalizationsEn();
-      // Reset IDs and cached controllers between tests
       mpLocator.mpGeneralController.reset();
     });
 
@@ -36,6 +34,7 @@ void main() {
         'file': '2025-10-04-001-area_and_line.th2',
         'length': 20,
         'encoding': 'UTF-8',
+        'lineID': 'blaus',
         'asFileOriginal': r'''encoding UTF-8
 scrap test
   area clay
@@ -69,6 +68,51 @@ scrap test
     3929.86 -568.03 3743.42 89.77 3592 208
   endline
   line contour -close on -id blaus -visibility off
+    2736.2 -808.5
+    2894.3 -202.7
+    2264.5 -205.7
+  endline
+endscrap
+''',
+      },
+      {
+        'file': '2025-10-05-002-area_and_line_without_id.th2',
+        'length': 20,
+        'encoding': 'UTF-8',
+        'lineID': null,
+        'asFileOriginal': r'''encoding UTF-8
+scrap test
+  area clay
+    l85-3732--20
+  endarea
+  line border -close on -id l85-3732--20 -visibility off
+    3592 208
+    3539.45 249.03 3447.39 245.1 3392 208
+    3233.22 101.65 3066.45 -131.93 3204 -332
+    3266.87 -423.45 3365.54 -513.28 3476 -524
+    3929.86 -568.03 3743.42 89.77 3592 208
+  endline
+  line contour
+    2736.2 -808.5
+    2894.3 -202.7
+    2264.5 -205.7
+  endline
+endscrap
+''',
+        'asFileChanged': r'''encoding UTF-8
+scrap test
+  area clay -id area1
+    l85-3732--20
+    area1-line1
+  endarea
+  line border -close on -id l85-3732--20 -visibility off
+    3592 208
+    3539.45 249.03 3447.39 245.1 3392 208
+    3233.22 101.65 3066.45 -131.93 3204 -332
+    3266.87 -423.45 3365.54 -513.28 3476 -524
+    3929.86 -568.03 3743.42 89.77 3592 208
+  endline
+  line contour -id area1-line1
     2736.2 -808.5
     2894.3 -202.7
     2264.5 -205.7
@@ -110,14 +154,20 @@ endscrap
             THLine? secondLine;
             for (final int mpID in parsedFile.linesMPIDs) {
               final THLine line = parsedFile.lineByMPID(mpID);
-
-              if (line.hasOption(THCommandOptionType.id) &&
-                  (line.optionByType(THCommandOptionType.id)
-                              as THIDCommandOption)
-                          .thID ==
-                      'blaus') {
-                secondLine = line;
-                break;
+              if (success['lineID'] == null) {
+                if (!line.hasOption(THCommandOptionType.id)) {
+                  secondLine = line;
+                  break;
+                }
+              } else {
+                if (line.hasOption(THCommandOptionType.id) &&
+                    (line.optionByType(THCommandOptionType.id)
+                                as THIDCommandOption)
+                            .thID ==
+                        'blaus') {
+                  secondLine = line;
+                  break;
+                }
               }
             }