+++ CHANGELOG.md
@@ -6,11 +6,13 @@
   * THFile creation through forCWJM constructor resulted in duplicated childrenMPIDs.
   * hash calculation and == operator not working in most elements with list or maps.
   * No implementation found for method getApplicationDocumentsDirectory.
+  * Original TH2FileLine being lost on set option.
 * Infrastructure maintenance:
   * Fixing error 'Binding has not yet been initialized' on all tests.
   * Removing pub test.
   * flutter pub upgrade --major-versions.
   * Implemented MPAddAreaTHIDCommand test.
+  * Adding add area THID test that creates area and line THIDs.
 
 ## 0.2.18 - 2025-09-26 - the anoying bug release
 * Fixed bugs: